feat(parser): separate parser and validator
